Moved name space of EnableOpenXRHandSettings
